% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{klfor2normals}
\alias{klfor2normals}
\title{klfor2normals
Compute the Kullback-Leibler divergence for 2 normal multivariate distributions}
\usage{
klfor2normals(theta1.mu, theta1.sigma, theta2.mu, theta2.sigma)
}
\arguments{
\item{theta1.mu}{the location parameter of the first distribution}

\item{theta1.sigma}{the covariance matrix of the first distribution}

\item{theta2.mu}{the location parameter of the second distribution}

\item{theta2.sigma}{the covariance matrix of the second distribution}
}
\value{
the K-L divergence.
}
\description{
klfor2normals
Compute the Kullback-Leibler divergence for 2 normal multivariate distributions
}
