% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad_disc.R
\name{quad_disc}
\alias{quad_disc}
\title{quad_disc}
\usage{
quad_disc(Y, theta.alpha, theta.mu, theta.sigma)
}
\arguments{
\item{Y}{A matrix of size n x p.}

\item{theta.alpha}{The  alpha values: An array of K positive real numbers
they must verify the condition sum(thetaOld.mu)== 1.}

\item{theta.mu}{The estimated centers: A list with K elements, each of them 
is an array of length p.}

\item{theta.sigma}{The estimated scatter matrices: A list with K matrices, each of them 
has dimension p x p}
}
\value{
A n x K matrix, where each row has the values of the quadratic discriminant 
with regarding to the j-th mixture component, j = 1,...,K
}
\description{
Computes the quadratic discriminant of each mixture component,
}
