% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustINIT.R
\name{robustINIT}
\alias{robustINIT}
\title{robustINIT}
\usage{
robustINIT(Y, K, nstart = 10)
}
\arguments{
\item{Y}{A matrix of size n x p.}

\item{K}{The number of groups}

\item{nstart}{the number of starting points to the algorithm, defaults to 10}
}
\value{
A list including the initial parameters of the mixture distribution, namely
\itemize{
\item{\code{alphaINIT}}{: K numeric values representing the convex combination coefficients.}
\item{\code{muINIT}}{: a list of length K with the location initial estimators.}
\item{\code{sigmaINIT}}{: a list of length K with the scatter matrix estimators.}
\item{\code{indicesINIT}}{: indices with initial clusters}
}
}
\description{
Robust Initializer for RMBC algorithm, it depends on the package \code{ktaucenters}
}
