% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sumkl}
\alias{sumkl}
\title{sumkl
The sum of K-L divergence measure between two successive iterations 
for each component of a mixture distribution,}
\usage{
sumkl(thetaNew.mu, thetaNew.sigma, thetaOld.mu, thetaOld.sigma)
}
\arguments{
\item{thetaNew.mu}{the location parameters of the first distribution}

\item{thetaNew.sigma}{the covariance matrix of the first distribution}

\item{thetaOld.mu}{the location parameter of the second distribution}

\item{thetaOld.sigma}{the covariance matrix of the second distribution}
}
\value{
the K-L divergence.
}
\description{
sumkl
The sum of K-L divergence measure between two successive iterations 
for each component of a mixture distribution,
}
