% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedMscale.R
\name{weightedMscale}
\alias{weightedMscale}
\title{weightedMscale
the M scale of an univariate sample (see reference below)}
\usage{
weightedMscale(u, b = 0.5, weights, c, initialsc = 0)
}
\arguments{
\item{u}{an univariate sample of size n.}

\item{b}{the desired break down point}

\item{weights}{the weights of each observation.}

\item{c}{a tuning constant, if consistency to standard normal distribution is desired use
\code{\link{normal_consistency_constants}}}

\item{initialsc}{the initial scale value, defaults to 0}
}
\value{
the weighted-Mscale value
}
\description{
weightedMscale
the M scale of an univariate sample (see reference below)
}
\references{
Maronna, R. A., Martin, R. D., Yohai, V. J., & Salibián-Barrera, M. (2018).
Robust statistics: theory and methods (with R). Wiley.
}
