% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedSestimator.R
\name{weightedSestimator}
\alias{weightedSestimator}
\title{weightedSestimator}
\usage{
weightedSestimator(
  Y,
  mu_init,
  sigma_init,
  max_iterFP = 1,
  weights,
  fixed_alpha
)
}
\arguments{
\item{Y}{A matrix of size n x p.}

\item{mu_init}{The previously computed center: an numerical array 
of length p.}

\item{sigma_init}{The previously computed scatter matrix: an array of numeric 
values p x p}

\item{max_iterFP}{the maximum number of fixed point iterations  used for the
algorithm, defaults to 1}

\item{weights}{The weights that contain the probability membership of each 
observation (related to the overall mixture components)}

\item{fixed_alpha}{the fixed alpha value for the corresponding mixture component}
}
\value{
A list including the estimated K centers and labels for the
    observations  
list(cov=matrixSigma,covAux1=covAux1,mu=muk,s=sk)
\itemize{
\item{\code{cov}}{:the computed weithted scatter matrix}
\item{\code{mu}}{: the computed weithted center}
\item{\code{s}}{: the weighted scale factor s.}
}
}
\description{
Computes the weighted location and scatter matrix estimators 
of the j-th mixture component , where the weights 
are calculated in the expectation-step.
}
