% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weightW}
\alias{weightW}
\title{weightW}
\usage{
weightW(arg, p)
}
\arguments{
\item{arg}{An 1-D array containing the distances.}

\item{p}{the dimension of the element}
}
\value{
an array of the same size of \code{arg} with the value of the weights
}
\description{
Weight function \code{ktaucenters}
}
